circle: hope to use python-libxdo when functional
The window type is expected to be "int" by libxdo. However, it is defined
as "ctypes.c_ulong" in python-libxdo. Changing "window_t" to
"ctypes.c_int" unforunately does not help, and further investigation in
that Python package would be needed to be able to use it, rather than
calling "xdotool".

Signed-off-by: Jimmy Durand Wesolowski <jimmy.wesolowski@esrlabs.com>